refactor(ld-preload): separated file access interceptors into standalone `libtermux-paths-ld-preload.so`
- moved `/etc/` path redirection interceptors (`open`, `openat`, `fopen`,
  `access`, `faccessat`, `stat`, `lstat`) from
  `TermuxExecDirectLDPreloadEntryPoint.c` into a new standalone entry point
  `TermuxPathsLDPreloadEntryPoint.c`
- reverted `TermuxExecDirectLDPreloadEntryPoint.c` to exec-only interception
- added `build-libtermux-paths-ld-preload` Makefile target with install,
  uninstall, format, and check support
